requirements.txt
certifi==2019.9.11
chardet==3.0.4
googlemaps==3.1.3
idna==2.8
numpy==1.17.3
ortools==7.4.7247
pandas==0.25.2
protobuf==3.10.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-rest-swagger==2.1.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
